json convert bug fixed

+++ Bums.cs
@@ -269,7 +269,7 @@ namespace BumsBot
         [JsonPropertyName("status")]
         public int Status { get; set; }
         [JsonPropertyName("nextLevelCost")]
-        public int NextLevelCost { get; set; }
+        public long NextLevelCost { get; set; }
         [JsonPropertyName("nextPerHourReward")]
         public string NextPerHourReward { get; set; } = string.Empty;
         [JsonPropertyName("limitText")]
